+++ test/codec.spec.ts
@@ -1,7 +1,7 @@
 /* eslint-env mocha */
 import * as codec from '../src/codec.js'
 import varint from 'varint'
-import { expect } from 'aegir/utils/chai.js'
+import { expect } from 'aegir/chai'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 
 describe('codec', () => {