+++ test/index.spec.ts
@@ -1,7 +1,7 @@
 /* eslint max-nested-callbacks: ["error", 8] */
 /* eslint-env mocha */
 import { Multiaddr } from '../src/index.js'
-import { expect } from 'aegir/utils/chai.js'
+import { expect } from 'aegir/chai'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import { codes } from '../src/protocols-table.js'
 
@@ -687,9 +687,49 @@ describe('helpers', () => {
 
     it('returns a node friendly address with dns', () => {
       expect(
+        new Multiaddr('/dns/wss0.bootstrap.libp2p.io/tcp/443').nodeAddress()
+      ).to.be.eql({
+        address: 'wss0.bootstrap.libp2p.io',
+        family: 4,
+        port: 443
+      })
+    })
+
+    it('returns a node friendly address with dns4', () => {
+      expect(
         new Multiaddr('/dns4/wss0.bootstrap.libp2p.io/tcp/443').nodeAddress()
       ).to.be.eql({
         address: 'wss0.bootstrap.libp2p.io',
+        family: 4,
+        port: 443
+      })
+    })
+
+    it('returns a node friendly address with dns6', () => {
+      expect(
+        new Multiaddr('/dns6/wss0.bootstrap.libp2p.io/tcp/443').nodeAddress()
+      ).to.be.eql({
+        address: 'wss0.bootstrap.libp2p.io',
+        family: 6,
+        port: 443
+      })
+    })
+
+    it('returns a node friendly address with dnsaddr', () => {
+      expect(
+        new Multiaddr('/dnsaddr/wss0.bootstrap.libp2p.io/tcp/443').nodeAddress()
+      ).to.be.eql({
+        address: 'wss0.bootstrap.libp2p.io',
+        family: 4,
+        port: 443
+      })
+    })
+
+    it('should transform a p2p dnsaddr without a tcp port into a node address', () => {
+      expect(
+        new Multiaddr('/dnsaddr/bootstrap.libp2p.io/p2p/QmNnooDu7bfjPFoTZYxMNLWUQJyrVwtbZg5gBMjTezGAJN').nodeAddress()
+      ).to.be.eql({
+        address: 'bootstrap.libp2p.io',
         family: 4,
         port: 443
       })